Fix cross-platform A64 CI failures

+++ src/buster/lib/compiler/assembly/aarch64_complex_simd_semantics.c
@@ -608,7 +608,10 @@ BusterA64SemanticVMValue buster_a64_complex_simd_value_arrangement(BusterA64Comp
     {
         return buster_a64_semantic_vm_value_invalid();
     }
-    return (BusterA64SemanticVMValue){.kind = BUSTER_A64_SEMANTIC_VM_VALUE_SIMD_ARRANGEMENT, .width = 8, .aux = arrangement, .payload = arrangement};
+    return (BusterA64SemanticVMValue){.kind = BUSTER_A64_SEMANTIC_VM_VALUE_SIMD_ARRANGEMENT,
+                                      .width = 8,
+                                      .aux = (u32)arrangement,
+                                      .payload = (u64)arrangement};
 }
 
 BusterA64SemanticVMValue buster_a64_complex_simd_value_register(u32 number, BusterA64ComplexSIMDArrangement arrangement, bool scalar)
@@ -620,7 +623,7 @@ BusterA64SemanticVMValue buster_a64_complex_simd_value_register(u32 number, Bust
     }
     return (BusterA64SemanticVMValue){.kind = scalar ? BUSTER_A64_SEMANTIC_VM_VALUE_SIMD_SCALAR : BUSTER_A64_SEMANTIC_VM_VALUE_SIMD_VECTOR,
                                       .width = width,
-                                      .aux = arrangement,
+                                      .aux = (u32)arrangement,
                                       .payload = number};
 }
 BusterA64SemanticVMValue buster_a64_complex_simd_value_vector(u32 number, BusterA64ComplexSIMDArrangement arrangement)
@@ -640,7 +643,11 @@ BusterA64SemanticVMValue buster_a64_complex_simd_value_list(u32 first, u32 count
     {
         return buster_a64_semantic_vm_value_invalid();
     }
-    return (BusterA64SemanticVMValue){.kind = BUSTER_A64_SEMANTIC_VM_VALUE_SIMD_LIST, .width = width, .aux = arrangement, .aux2 = count, .payload = first};
+    return (BusterA64SemanticVMValue){.kind = BUSTER_A64_SEMANTIC_VM_VALUE_SIMD_LIST,
+                                      .width = width,
+                                      .aux = (u32)arrangement,
+                                      .aux2 = count,
+                                      .payload = first};
 }
 
 BusterA64SemanticVMValue buster_a64_complex_simd_value_lane(u32 number, BusterA64ComplexSIMDArrangement arrangement, u32 lane)
@@ -651,7 +658,11 @@ BusterA64SemanticVMValue buster_a64_complex_simd_value_lane(u32 number, BusterA6
     {
         return buster_a64_semantic_vm_value_invalid();
     }
-    return (BusterA64SemanticVMValue){.kind = BUSTER_A64_SEMANTIC_VM_VALUE_SIMD_LANE, .width = width, .aux = arrangement, .aux2 = lane, .payload = number};
+    return (BusterA64SemanticVMValue){.kind = BUSTER_A64_SEMANTIC_VM_VALUE_SIMD_LANE,
+                                      .width = width,
+                                      .aux = (u32)arrangement,
+                                      .aux2 = lane,
+                                      .payload = number};
 }
 BusterA64SemanticVMValue buster_a64_complex_simd_value_gpr(u32 number, u8 width, bool zr)
 {
@@ -1366,7 +1377,10 @@ static bool buster_a64_complex_simd_decode_register(u32 row_index, BusterA64Sema
         {
             return false;
         }
-        *result = (BusterA64SemanticVMValue){.kind = BUSTER_A64_SEMANTIC_VM_VALUE_SIMD_VECTOR, .width = width, .aux = arrangement, .payload = number};
+        *result = (BusterA64SemanticVMValue){.kind = BUSTER_A64_SEMANTIC_VM_VALUE_SIMD_VECTOR,
+                                             .width = width,
+                                             .aux = (u32)arrangement,
+                                             .payload = number};
     }
     else
     {
